Move PHPCR-ODM adapter to package
