+++ lib/Adapter/Doctrine/PHPCRODM/QueryAdapter.php
@@ -1,0 +1,56 @@
+<?php
+
+namespace Pagerfanta\Doctrine\PHPCRODM;
+
+use Doctrine\ODM\PHPCR\Query\Builder\QueryBuilder;
+use Doctrine\ODM\PHPCR\Query\Query;
+use Pagerfanta\Adapter\AdapterInterface;
+
+/**
+ * Adapter which calculates pagination from a Doctrine PHPCR ODM QueryBuilder.
+ */
+class QueryAdapter implements AdapterInterface
+{
+    /**
+     * @var QueryBuilder
+     */
+    private $queryBuilder;
+
+    public function __construct(QueryBuilder $queryBuilder)
+    {
+        $this->queryBuilder = $queryBuilder;
+    }
+
+    /**
+     * @return QueryBuilder
+     */
+    public function getQueryBuilder()
+    {
+        return $this->queryBuilder;
+    }
+
+    /**
+     * @return int
+     */
+    public function getNbResults()
+    {
+        return $this->queryBuilder->getQuery()
+            ->execute(null, Query::HYDRATE_PHPCR)
+            ->getRows()
+            ->count();
+    }
+
+    /**
+     * @param int $offset
+     * @param int $length
+     *
+     * @return iterable
+     */
+    public function getSlice($offset, $length)
+    {
+        return $this->queryBuilder->getQuery()
+            ->setMaxResults($length)
+            ->setFirstResult($offset)
+            ->execute();
+    }
+}
